[wip] sparse case
