+++ src/asgard_kronmult_matrix.cpp
@@ -331,197 +331,6 @@ inline bool check_connected(int const num_dimensions, int const *const row,
   return true;
 }
 
-/*!
- * \brief Keeps track of the connectivity of the elements in the 1d hierarchy.
- *
- * Constructs a sparse matrix-like structure with row-compressed format and
- * ordered indexes within each row, so that the 1D connectivity can be verified
- * with a simple binary search.
- *
- * The advantage of the structure is to provide:
- * - easier check if two 1D cell are connected or not
- * - index of the connection, so operator coefficient matrices can be easily
- *   referenced
- *
- * TODO: This logic should be moved inside the coefficients class.
- */
-class connect_1d
-{
-public:
-  /*!
-   *  \brief Constructor, makes the connectivity up to and including the given
-   *         max-level.
-   */
-  connect_1d(int const max_level)
-      : levels(max_level), cells(1 << levels), pntr(cells + 1, 0),
-        indx(2 * cells)
-  {
-    std::vector<int> cell_per_level(levels + 2, 1);
-    for (int l = 2; l < levels + 2; l++)
-      cell_per_level[l] = 2 * cell_per_level[l - 1];
-
-    // first two cells are connected to everything
-    pntr[1] = cells;
-    pntr[2] = 2 * cells;
-    for (int i = 0; i < cells; i++)
-      indx[i] = i;
-    for (int i = 0; i < cells; i++)
-      indx[i + cells] = i;
-
-    // for the remaining, loop level by level, cell by cell
-    for (int l = 2; l < levels + 1; l++)
-    {
-      int level_size = cell_per_level[l]; // number of cells in this level
-
-      // for each cell in this level, look at all cells connected
-      // look at previous levels, this level, follow on levels
-
-      // start with the first cell, on the left edge
-      int i = level_size; // index of the first cell
-      // always connected to cells 0 and 1
-      indx.push_back(0);
-      indx.push_back(1);
-      // look at cells above
-      for (int upl = 2; upl < l; upl++)
-      {
-        // edge cell is connected to both edge cells on each level
-        indx.push_back(cell_per_level[upl]);
-        indx.push_back(cell_per_level[upl + 1] - 1);
-      }
-      // look at this level
-      indx.push_back(i);
-      indx.push_back(i + 1);
-      // connect also to the right-most cell (periodic boundary)
-      if (l > 2) // at level l = 2, i+1 is the right-most cell
-        indx.push_back(cell_per_level[l + 1] - 1);
-      // look at follow on levels
-      for (int downl = l + 1; downl < levels + 1; downl++)
-      {
-        // connect to the first bunch of cell, i.e., with overlapping support
-        // going on by 2, 4, 8 ... and one more for touching boundary
-        // also connect to the right-most cell
-        int lstart = cell_per_level[downl];
-        for (int downp = 0; downp < cell_per_level[downl - l + 1] + 1; downp++)
-          indx.push_back(lstart + downp);
-        indx.push_back(cell_per_level[downl + 1] - 1);
-      }
-      pntr[i + 1] = static_cast<int>(indx.size()); // done with point
-
-      // handle middle cells
-      for (int p = 1; p < level_size - 1; p++)
-      {
-        i++;
-        // always connected to the first two cells
-        indx.push_back(0);
-        indx.push_back(1);
-        // ancestors on previous levels
-        for (int upl = 2; upl < l; upl++)
-        {
-          int segment_size = cell_per_level[l - upl + 1];
-          int ancestor     = p / segment_size;
-          int edge         = p - ancestor * segment_size; // p % segment_size
-          // if on the left edge of the ancestor
-          if (edge == 0)
-            indx.push_back(cell_per_level[upl] + ancestor - 1);
-          indx.push_back(cell_per_level[upl] + ancestor);
-          // if on the right edge of the ancestor
-          if (edge == segment_size - 1)
-            indx.push_back(cell_per_level[upl] + ancestor + 1);
-        }
-        // on this level
-        indx.push_back(i - 1);
-        indx.push_back(i);
-        indx.push_back(i + 1);
-        // kids on further levels
-        int left_kid = p; // initialize, will be updated on first iteration
-        int num_kids = 1;
-        for (int downl = l + 1; downl < levels + 1; downl++)
-        {
-          left_kid *= 2;
-          num_kids *= 2;
-          for (int j = left_kid - 1; j < left_kid + num_kids + 1; j++)
-            indx.push_back(cell_per_level[downl] + j);
-        }
-        pntr[i + 1] = static_cast<int>(indx.size()); // done with cell i
-      }
-
-      // right edge cell
-      i++;
-      // always connected to 0 and 1
-      indx.push_back(0);
-      indx.push_back(1);
-      for (int upl = 2; upl < l; upl++)
-      {
-        // edge cell is connected to both edge cells on each level
-        indx.push_back(cell_per_level[upl]);
-        indx.push_back(cell_per_level[upl + 1] - 1);
-      }
-      // at this level
-      // connect also to the left-most cell (periodic boundary)
-      if (l > 2) // at level l = 2, left-most cell is i-1, don't double add
-        indx.push_back(cell_per_level[l]);
-      indx.push_back(i - 1);
-      indx.push_back(i);
-      // look at follow on levels
-      for (int downl = l + 1; downl < levels + 1; downl++)
-      {
-        // left edge on the level
-        indx.push_back(cell_per_level[downl]);
-        // get the last bunch of cells at the level
-        int lend = cell_per_level[downl + 1] - 1;
-        for (int downp = cell_per_level[downl - l + 1]; downp > -1; downp--)
-          indx.push_back(lend - downp);
-      }
-      pntr[i + 1] = static_cast<int>(indx.size()); // done with the right edge
-    } // done with level, move to the next level
-  }   // close the constructor
-
-  int get_offset(int row, int col) const
-  {
-    // first two levels are large and trivial, no need to search
-    if (row == 0)
-      return col;
-    else if (row == 1)
-      return cells + col;
-    // if not on the first or second row, do binary search
-    int sstart = pntr[row], send = pntr[row + 1] - 1;
-    int current = (sstart + send) / 2;
-    while (sstart <= send)
-    {
-      if (indx[current] < col)
-      {
-        sstart = current + 1;
-      }
-      else if (indx[current] > col)
-      {
-        send = current - 1;
-      }
-      else
-      {
-        return current;
-      };
-      current = (sstart + send) / 2;
-    }
-    return -1;
-  }
-
-  int num_connections() const { return static_cast<int>(indx.size()); }
-
-  int num_cells() const { return cells; }
-
-  int row_begin(int row) const { return pntr[row]; }
-
-  int row_end(int row) const { return pntr[row + 1]; }
-
-  int operator[](int j) const { return indx[j]; }
-
-private:
-  int levels;
-  int cells;
-  std::vector<int> pntr;
-  std::vector<int> indx;
-};
-
 template<typename precision>
 kronmult_matrix<precision>
 make_kronmult_sparse(PDE<precision> const &pde,
@@ -831,7 +640,8 @@ void update_kronmult_coefficients(PDE<P> const &pde,
 template<typename P>
 memory_usage
 compute_mem_usage(PDE<P> const &pde, adapt::distributed_grid<P> const &discretization,
-                  options const &program_options, imex_flag const imex)
+                  options const &program_options, imex_flag const imex,
+                  kron_sparse_cache &spcache)
 {
   auto const &grid         = discretization.get_subgrid(get_rank());
   int const num_dimensions = pde.num_dims;
@@ -928,7 +738,98 @@ compute_mem_usage(PDE<P> const &pde, adapt::distributed_grid<P> const &discretiz
   }
   else
   { // sparse mode
-      //
+    // if possible, keep the 1d connectivity matrix
+    if (pde.max_level != spcache.cells1d.max_loaded_level())
+      spcache.cells1d = connect_1d(pde.max_level);
+
+    int const *const flattened_table =
+        discretization.get_table().get_active_table().data();
+
+    // This is a bad algorithm as it loops over all possible pairs of
+    // multi-indexes The correct algorithm is to infer the connectivity from the
+    // sparse grid graph hierarchy and avoid doing so many comparisons, but that
+    // requires messy work with the way the indexes are stored in memory. To do
+    // this properly, I need a fast map from a multi-index to the matrix row
+    // associated with the multi-index (or indicate if it's missing).
+    // The unordered_map does not provide this functionality and the addition of
+    // the flattened table in element.hpp is not a good answer.
+    // Will fix in a future PR ...
+
+    if (spcache.ccount.size() == num_rows)
+      std::fill(spcache.begin(), spcache.end(), 0);
+    else
+      spcache.ccount = std::vector<int>(num_rows);
+
+#pragma omp parallel for
+    for (int64_t row = grid.row_start; row < grid.row_stop + 1; row++)
+    {
+      int const *const row_coords = flattened_table + 2 * num_dimensions * row;
+      // (L, p) = (row_coords[i], row_coords[i + num_dimensions])
+      for (int64_t col = grid.col_start; col < grid.col_stop + 1; col++)
+      {
+        int const *const col_coords = flattened_table + 2 * num_dimensions * col;
+        if (check_connected(num_dimensions, row_coords, col_coords))
+          spcache.ccount[row - grid.row_start]++;
+      }
+    }
+
+    spcache.num_nonz = 0; // total number of connected cells
+    for(auto const c : spcache.ccount)
+      spcache.num_nonz += c;
+
+    int64_t base_line_entries = stats.cells1d.num_connections() * num_dimensions * num_terms * kron_size * kron_size;
+    stats.baseline_memory = get_MB<P>(base_line_entries);
+
+#ifdef ASGARD_USE_CUDA
+    int64_t available_MB = program_options.memory_limit - stats.baseline_memory;
+    std::cout << " available_MB = " << available_MB << "\n";
+    check_available_memory(stats.baseline_memory, available_MB);
+    std::cout << " stats.baseline_memory = " << stats.baseline_memory << "\n";
+
+    // 2147483646 = 2^31 - 2, which prevents overflow in the 32-bit signed int
+    int64_t available_entries = std::min(int64_t{2147483646}, (available_MB * 1024 * 1024) / static_cast<int64_t>(sizeof(int)));
+#else
+    // CPU mode is limited only by the 32-bit indexing
+    int64_t available_entries = 2147483646;
+#endif
+
+    // available_entries holds the number of ints that can be loaded at
+    // one time to respect the memory limit
+    if (imex == imex_flag::unspecified)
+    {
+      // assuming there will be only one matrix, compute the number of indexes
+      // each product uses pde.num_terms * num_dimensions ints
+      // plus two more for the row and column index
+      int64_t size_of_indexes =
+          spcache.num_nonz * (pde.num_terms * num_dimensions + 2);
+
+      if (size_of_indexes <= available_entries)
+      {
+        stats.kron_call = memory_usage::one_call;
+      }
+      else
+      {
+        stats.kron_call = memory_usage::multi_calls;
+        if (available_entries == 2147483646)
+        {
+          stats.mem_limit = memory_usage::overflow;
+#ifdef ASGARD_USE_CUDA
+          stats.work_size = available_entries / 2;
+#else
+          stats.work_size = available_entries;
+#endif
+        }
+        else
+        {
+          stats.mem_limit = memory_usage::environment;
+        }
+      }
+    }
+    else
+    {
+      // assuming there will be two matrices, find the memory for each
+    }
+
   }
 
   return stats;
@@ -947,7 +848,7 @@ template void update_kronmult_coefficients<double>(PDE<double> const &,
 template memory_usage
 compute_mem_usage<double>(PDE<double> const &,
                           adapt::distributed_grid<double> const &,
-                          options const &, imex_flag const);
+                          options const &, imex_flag const, kron_sparse_cache &);
 #endif
 
 #ifdef ASGARD_ENABLE_FLOAT
@@ -962,7 +863,7 @@ update_kronmult_coefficients<float>(PDE<float> const &, options const &,
 template memory_usage
 compute_mem_usage<float>(PDE<float> const &,
                          adapt::distributed_grid<float> const &,
-                         options const &, imex_flag const);
+                         options const &, imex_flag const, kron_sparse_cache &);
 #endif
 
 } // namespace asgard
